task_maker_lang

Struct SourceFile

pub struct SourceFile {
    pub path: PathBuf,
    pub base_path: PathBuf,
    pub language: Arc<dyn Language>,
    pub executable: Arc<Mutex<Option<File>>>,
    pub grader_map: Option<Arc<GraderMap>>,
    pub copy_exe: bool,
    pub write_bin_to: Option<PathBuf>,
    pub link_static: bool,
}

A source file that will be able to be executed (with an optional compilation step).

After creating a SourceFile using new you can add start using it via the execute method. Note that it may add to the DAG an extra execution for compiling the source file.

Fields
path: PathBuf

Path to the source file.

base_path: PathBuf

Path to the base directory (e.g. the task root), used for including the source file dependencies from the command line args of the executable (in case of relative paths).

language: Arc<dyn Language>

Language of the source file.

executable: Arc<Mutex<Option<File>>>

Handle to the executable after the compilation/provided file.

grader_map: Option<Arc<GraderMap>>

An optional handler to the map of the graders.

copy_exe: bool

Whether to force the copy-exe option of the DAG for this source file.

write_bin_to: Option<PathBuf>

Where to write the compiled executable.

link_static: bool

Whether this source file should be statically linked.

Implementations

impl SourceFile

pub fn new<P: Into<PathBuf>, P2: Into<PathBuf>, P3: Into<PathBuf>>( path: P, base_path: P2, grader_map: Option<Arc<GraderMap>>, write_bin_to: Option<P3>, ) -> Option<SourceFile>

Make a new SourceFile from the provided file. Will return None if the language is unknown.

The language of the source file will be detected using the LanguageManager, only those languages are supported.

Because the execution/compilation may require some additional files a GraderMap is required.

pub fn execute<S: AsRef<str>>( &self, dag: &mut ExecutionDAG, description: S, args: Vec<String>, ) -> Result<(Option<ExecutionUuid>, Execution), Error>

Execute the program relative to this source file with the specified args. If the file has not been compiled yet this may add the compilation to the DAG. The compilation is added to the DAG only once for each SourceFile instance.

The returned execution has the following properties already set:

the list of arguments

the input file for the executable

the input files from the language runtime dependencies

the input files from the grader map runtime dependencies

The first element returned is the UUID of the execution of the compilation. It will be returned only once, even if the execute method is called more than once. The actual execution is returned as second parameter.

The returned execution has all the dependencies already set, but it has not been added to the DAG yet. In order for this execution to work it has to be manually added to the DAG using ExecutionDAG::add_execution.

Examples

When executing a .cpp file the first item returned contains an handle to the compilation execution. Note that the second time the handle is not returned.

use task_maker_dag::ExecutionDAG;
use task_maker_lang::SourceFile;

let mut dag = ExecutionDAG::new();
let mut source = SourceFile::new(path /* test.cpp */, "", None, None::<PathBuf>).unwrap();

let (comp, exec) = source.execute(&mut dag, "Execution", vec!["arg1".into()]).unwrap();
assert!(comp.is_some());
// customize the execution...
dag.add_execution(exec);

let (comp, exec) = source.execute(&mut dag, "Execution 2", vec!["arg1".into()]).unwrap();
assert!(comp.is_none());
dag.add_execution(exec);

When executing a .py file the handle is not returned.

use task_maker_dag::ExecutionDAG;
use task_maker_lang::SourceFile;

let mut dag = ExecutionDAG::new();
let mut source = SourceFile::new(path /* test.py */, "", None, None::<PathBuf>).unwrap();

let (comp, exec) = source.execute(&mut dag, "Execution", vec!["arg1".into()]).unwrap();
assert!(comp.is_none());
// customize the execution...
dag.add_execution(exec);

pub fn copy_exe(&mut self)

Force the executable to be copied to write_bin_to regardless of the option of the DAG.

pub fn link_static(&mut self)

Compile the source file to a statically-linked binary.

pub fn executable( &self, dag: &mut ExecutionDAG, ) -> Result<(FileUuid, Option<ExecutionUuid>), Error>

Prepare the source file if needed and return the executable file. If the compilation step was not executed yet the handle to the compilation execution is also returned.

pub fn name(&self) -> String

The file name of the source file.

use task_maker_lang::SourceFile;
use std::path::PathBuf;

let source = SourceFile::new("path/to/sourcefile.cpp", "", None, None::<PathBuf>).unwrap();

assert_eq!(source.name(), "sourcefile.cpp");

pub fn relative_path(&self) -> &Path

The path to the file, relative to the base_dir if possible. If the file is not inside the base dir, the full path is returned.

pub fn write_bin_to(&self) -> Option<PathBuf>

The optional destination of where to copy the executable if copy-exe option is set.

use task_maker_lang::SourceFile;

let source = SourceFile::new("path/to/sourcefile.cpp", "", None, Some("exec")).unwrap();

assert_eq!(source.write_bin_to(), Some("exec".into()));

pub fn prepare( &self, dag: &mut ExecutionDAG, ) -> Result<Option<ExecutionUuid>, Error>

Prepare the source file setting the executable and eventually compiling the source file.

pub fn language(&self) -> &dyn Language

The language of the source file.