Crate task_maker_lang

Crate for managing programming languages and source files.

This crate will expose some utility functions for making a DAG with the execution of commands that come directly from source files.

The Language trait exposes the interface for defining new programming languages. The list of supported programming languages can be found in the source of this crate.

The entry point of this crate is LanguageManager, a struct that is able to detect the language of a source file based on its extension. A trait object is used to keep track of the language.

To actually use the language you can use SourceFile, it exposes the functionalities for compiling and running a source file.

Example

use task_maker_lang::LanguageManager;

let lang = LanguageManager::detect_language("test.cpp").expect("unknown lang");
assert!(lang.name().contains("C++"))

Structs

Dependency

A dependency of an execution, all the sandbox paths must be relative and inside of the sandbox.

GraderMap

The storage of the compilation/runtime dependencies for the source files.

LanguageManager

Manager of all the known languages, you should use this to get Language instances.

SourceFile

A source file that will be able to be executed (with an optional compilation step).

Traits

Language

Trait that defines the properties of the supported languages. Most of the methods have a safe blanket implementation, note that not all of them are really optional: based on the value returned by need_compilation some of the methods become required.