task_maker_lang
Struct GraderMap

pub struct GraderMap { /* private fields */ }

The storage of the compilation/runtime dependencies for the source files.

A source file may need some extra dependencies in order to be compiled and/or executed. For example a C++ file may need a second C++ file to be linked together, or a Python file may need a second Python file to be run.

Implementations

impl GraderMap

pub fn new<P: Into<PathBuf>>(graders: Vec<P>) -> GraderMap

Make a new map with the specified graders.

use task_maker_lang::GraderMap;

let map = GraderMap::new(vec!["file.cpp", "file.py"]);

pub fn get_compilation_deps(&self, lang: &dyn Language) -> Option<Dependency>

Extra compilation dependencies of the graders, will be an empty Vec if the language is not compiled.

use task_maker_lang::{GraderMap, LanguageManager};

let map = GraderMap::new(vec!["file.cpp", "file.py"]);
let cpp = LanguageManager::detect_language("source.cpp").unwrap();
let py = LanguageManager::detect_language("source.py").unwrap();
assert!(map.get_compilation_deps(cpp.as_ref()).is_some());
assert!(map.get_compilation_deps(py.as_ref()).is_none());

pub fn get_runtime_deps(&self, lang: &dyn Language) -> Vec<Dependency>

Extra runtime dependencies of the graders, will be an empty Vec if the language is compiled.

use task_maker_lang::{GraderMap, LanguageManager};

let map = GraderMap::new(vec!["file.cpp", "file.py"]);
let cpp = LanguageManager::detect_language("source.cpp").unwrap();
let py = LanguageManager::detect_language("source.py").unwrap();
assert_eq!(map.get_runtime_deps(cpp.as_ref()).len(), 0);
assert_eq!(map.get_runtime_deps(py.as_ref()).len(), 1);

pub fn all_paths(&self) -> impl Iterator<Item = &Path>

Return an iterator over the paths of all the graders in this map.

use task_maker_lang::{GraderMap, LanguageManager};
use std::path::Path;

let map = GraderMap::new(vec!["file.cpp"]);
assert_eq!(map.all_paths().collect::<Vec<_>>(), vec![Path::new("file.cpp")]);