Trait task_maker_lang::Language

pub trait Language: Debug + Send + Sync {

    fn name(&self) -> &'static str;

    fn extensions(&self) -> Vec<&'static str>;

    fn need_compilation(&self) -> bool;

    fn compilation_command(
        &self, 
        _path: &Path, 
        _write_to: Option<&Path>
    ) -> ExecutionCommand { ... }

    fn compilation_args(
        &self, 
        _path: &Path, 
        _write_to: Option<&Path>, 
        _link_static: bool
    ) -> Vec<String> { ... }

    fn compilation_add_file(
        &self, 
        _args: Vec<String>, 
        _file: &Path
    ) -> Vec<String> { ... }

    fn compilation_dependencies(&self, _path: &Path) -> Vec<Dependency> { ... }

    fn compiled_file_name(
        &self, 
        _path: &Path, 
        _write_to: Option<&Path>
    ) -> PathBuf { ... }

    fn runtime_command(
        &self, 
        path: &Path, 
        write_to: Option<&Path>
    ) -> ExecutionCommand { ... }

    fn runtime_args(
        &self, 
        _path: &Path, 
        _write_to: Option<&Path>, 
        args: Vec<String>
    ) -> Vec<String> { ... }

    fn runtime_add_file(&self, args: Vec<String>, _file: &Path) -> Vec<String> { ... }

    fn runtime_dependencies(&self, _path: &Path) -> Vec<Dependency> { ... }

    fn custom_limits(&self, _limits: &mut ExecutionLimits) { ... }

    fn executable_name(&self, path: &Path, write_to: Option<&Path>) -> PathBuf { ... }

}

Trait that defines the properties of the supported languages. Most of the methods have a safe blanket implementation, note that not all of them are really optional: based on the value returned by need_compilation some of the methods become required.

A language can be either compiler or not-compiler.

When a language is compiled the extra required implementations are:

compilation_command

compilation_args

compilation_add_file

Required methods

fn name(&self) -> &'static str

Full name of the language. This must be unique between all the other languages.

fn extensions(&self) -> Vec<&'static str>

List of valid extensions for this language. A file is considered in this language if its extension is inside this list.

fn need_compilation(&self) -> bool

Whether this language needs a compilation step. Returning true here triggers many changes in the behaviour of the execution. Of course the compilation step will be added, because of that there is the need to know how to compile the source file, forcing the implementation of some extra methods.

Provided methods

fn compilation_command(
&self,
_path: &Path,
_write_to: Option<&Path>
) -> ExecutionCommand

Command to use to compile the source file. The blanked implementation is intended for not compiled languages.

Will panic if this language does not support compilation.

fn compilation_args(
&self,
_path: &Path,
_write_to: Option<&Path>,
_link_static: bool
) -> Vec<String>

Arguments to pass to the compiler to compile to source file. The source file is located at path.file_name() inside the sandbox and the result of the compilation should placed at self.compiled_file_name(path). The blanked implementation is intended for not compiled languages.

Will panic if this language does not support compilation.

fn compilation_add_file(&self, _args: Vec<String>, _file: &Path) -> Vec<String>

Add a file to the compilation command if the language requires that. That file can be any compile time dependency and it’s relative to the sandbox.

The new compilation arguments should be returned.

Will panic if this language does not support compilation.

fn compilation_dependencies(&self, _path: &Path) -> Vec<Dependency>

The dependencies to put inside the compilation sandbox. This does not include the source file.

fn compiled_file_name(&self, _path: &Path, _write_to: Option<&Path>) -> PathBuf

The name that the compiled file will have inside the compilation sandbox.

fn runtime_command(
&self,
path: &Path,
write_to: Option<&Path>
) -> ExecutionCommand

Command to use to run the program. It defaults to the executable name of the program. Languages that need to run a separate program (e.g. a system-wise interpreter) may change the return value of this method.

fn runtime_args(
&self,
_path: &Path,
_write_to: Option<&Path>,
args: Vec<String>
) -> Vec<String>

Arguments to pass to the executable to start the evaluation.

fn runtime_add_file(&self, args: Vec<String>, _file: &Path) -> Vec<String>

Add a file to the runtime command if the language requires that. That file can be any run time dependency and it’s relative to the sandbox.

The new runtime arguments should be returned.

fn runtime_dependencies(&self, _path: &Path) -> Vec<Dependency>

The dependencies to put inside the execution sandbox. This does not include the executable.

fn custom_limits(&self, _limits: &mut ExecutionLimits)

Update the limits for some language-specific requirements. For example the executable may need to fork (hence use more processes).

fn executable_name(&self, path: &Path, write_to: Option<&Path>) -> PathBuf

The name of the executable inside the sandbox. If this binary will be written elsewhere in the system, use the same name. Otherwise fallback to the original file name, without extension.